Cleanup and Improve mul/div with/by +1 and -1

+++ test/test.cpp
@@ -1,5 +1,5 @@
 ﻿///////////////////////////////////////////////////////////////////
-//  Copyright Christopher Kormanyos 1999 - 2022.                 //
+//  Copyright Christopher Kormanyos 1999 - 2023.                 //
 //  Distributed under the Boost Software License,                //
 //  Version 1.0. (See accompanying file LICENSE_1_0.txt          //
 //  or copy at http://www.boost.org/LICENSE_1_0.txt)             //
@@ -92,7 +92,7 @@ auto main() -> int
   const auto stop = local_clock_type::now();
 
   {
-    constexpr auto one_thousand_milliseconds = 1000.0F;
+    constexpr auto one_thousand_milliseconds = static_cast<float>(1000.0F);
 
     const auto test_duration_sec =
       static_cast<float>
